Fix macro redefinition warning when including <inttypes.h> after bson.h using Visual Studio 2013 or newer. Fix iov_len in mongoc_iovec_t on Windows to be size_t like on Linux

+++ src/libmongoc/src/mongoc/mongoc-iovec.h
@@ -32,7 +32,7 @@ BSON_BEGIN_DECLS
 
 #ifdef _WIN32
 typedef struct {
-   u_long iov_len;
+   size_t iov_len;
    char *iov_base;
 } mongoc_iovec_t;
 
@@ -40,10 +40,10 @@ BSON_STATIC_ASSERT2 (sizeof_iovect_t,
                      sizeof (mongoc_iovec_t) == sizeof (WSABUF));
 BSON_STATIC_ASSERT2 (offsetof_iovec_base,
                      offsetof (mongoc_iovec_t, iov_base) ==
-                     offsetof (WSABUF, buf));
+                        offsetof (WSABUF, buf));
 BSON_STATIC_ASSERT2 (offsetof_iovec_len,
                      offsetof (mongoc_iovec_t, iov_len) ==
-                     offsetof (WSABUF, len));
+                        offsetof (WSABUF, len));
 
 #else
 typedef struct iovec mongoc_iovec_t;
